+++ providers/jaxb/src/main/java/org/jboss/resteasy/plugins/providers/jaxb/MapProvider.java
@@ -35,7 +35,7 @@ import javax.xml.transform.stream.StreamSource;
 import org.jboss.resteasy.annotations.providers.jaxb.DoNotUseJAXBProvider;
 import org.jboss.resteasy.annotations.providers.jaxb.WrappedMap;
 import org.jboss.resteasy.core.ResteasyContext;
-import org.jboss.resteasy.core.interception.jaxrs.AsyncBufferedMessageBodyWriter;
+import org.jboss.resteasy.core.messagebody.AsyncBufferedMessageBodyWriter;
 import org.jboss.resteasy.plugins.providers.jaxb.i18n.LogMessages;
 import org.jboss.resteasy.plugins.providers.jaxb.i18n.Messages;
 import org.jboss.resteasy.spi.ResteasyConfiguration;
