+++ providers/json-p-ee7/src/main/java/org/jboss/resteasy/plugins/providers/jsonp/JsonValueProvider.java
@@ -18,7 +18,7 @@ import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.MultivaluedMap;
 import javax.ws.rs.ext.MessageBodyReader;
 
-import org.jboss.resteasy.core.interception.jaxrs.AsyncBufferedMessageBodyWriter;
+import org.jboss.resteasy.core.messagebody.AsyncBufferedMessageBodyWriter;
 import org.jboss.resteasy.resteasy_jaxrs.i18n.LogMessages;
 /**
  * @author <a href="mailto:ema@redhat.com">Jim Ma</a>
