+++ resteasy-core-spi/src/main/java/org/jboss/resteasy/spi/AsyncMessageBodyWriter.java
@@ -1,4 +1,4 @@
-package org.jboss.resteasy.core.interception.jaxrs;
+package org.jboss.resteasy.spi;
 
 import java.lang.annotation.Annotation;
 import java.lang.reflect.Type;
@@ -7,8 +7,6 @@ import java.util.concurrent.CompletionStage;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.MultivaluedMap;
 import javax.ws.rs.ext.MessageBodyWriter;
-
-import org.jboss.resteasy.spi.AsyncOutputStream;
 
 public interface AsyncMessageBodyWriter<T> extends MessageBodyWriter<T> {
     CompletionStage<Void> asyncWriteTo(T t, Class<?> type, Type genericType, Annotation[] annotations, MediaType mediaType,
