+++ resteasy-core/src/main/java/org/jboss/resteasy/plugins/providers/ByteArrayProvider.java
@@ -1,7 +1,7 @@
 package org.jboss.resteasy.plugins.providers;
 
-import org.jboss.resteasy.core.interception.jaxrs.AsyncMessageBodyWriter;
 import org.jboss.resteasy.resteasy_jaxrs.i18n.LogMessages;
+import org.jboss.resteasy.spi.AsyncMessageBodyWriter;
 import org.jboss.resteasy.spi.AsyncOutputStream;
 import org.jboss.resteasy.util.MediaTypeHelper;
 import org.jboss.resteasy.util.NoContent;
