+++ resteasy-core/src/main/java/org/jboss/resteasy/plugins/providers/DefaultTextPlain.java
@@ -1,6 +1,6 @@
 package org.jboss.resteasy.plugins.providers;
 
-import org.jboss.resteasy.core.interception.jaxrs.AsyncMessageBodyWriter;
+import org.jboss.resteasy.spi.AsyncMessageBodyWriter;
 import org.jboss.resteasy.spi.AsyncOutputStream;
 import org.jboss.resteasy.util.NoContent;
 import org.jboss.resteasy.util.TypeConverter;
