+++ resteasy-core/src/main/java/org/jboss/resteasy/plugins/providers/FormUrlEncodedProvider.java
@@ -1,9 +1,9 @@
 package org.jboss.resteasy.plugins.providers;
 
-import org.jboss.resteasy.core.interception.jaxrs.AsyncMessageBodyWriter;
 import org.jboss.resteasy.resteasy_jaxrs.i18n.LogMessages;
 import org.jboss.resteasy.specimpl.MultivaluedMapImpl;
 import org.jboss.resteasy.util.Encode;
+import org.jboss.resteasy.spi.AsyncMessageBodyWriter;
 import org.jboss.resteasy.spi.AsyncOutputStream;
 import org.jboss.resteasy.spi.util.FindAnnotation;
 import org.jboss.resteasy.util.NoContent;
