+++ resteasy-core/src/main/java/org/jboss/resteasy/plugins/providers/SourceProvider.java
@@ -1,6 +1,6 @@
 package org.jboss.resteasy.plugins.providers;
 
-import org.jboss.resteasy.core.interception.jaxrs.AsyncBufferedMessageBodyWriter;
+import org.jboss.resteasy.core.messagebody.AsyncBufferedMessageBodyWriter;
 import org.jboss.resteasy.resteasy_jaxrs.i18n.LogMessages;
 import org.jboss.resteasy.util.NoContent;
 import org.xml.sax.InputSource;
